requirements.txt
APScheduler==3.6.3
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-apscheduler==0.3.0
django-filter==2.3.0
django-filters==0.2.1
djangorestframework==3.11.0
idna==2.10
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.9
